refactor: split analysis/prompts/builder.py into focused modules
